Added detailed introduction and minor code changes
Have added a more detailed information portion at the beginning of the
code in the form of comments. Have made minor changes to the code.
Specifically, have changed the 'docker' user's login shell from
'/bin/sh' to '/bin/ash', which is the default shell in Alpine. Have also
added a line that updates the repositories using the package manager.