Improved Dockerfile using suggested practices
Have grouped the previous 3 RUN commands into one RUN command. RUN
will execute any commands in a new layer on top of the current image and
commit the results. The resulting committed image will be used for the
next step in the Dockerfile. Therefore, it is recommended that all
commands to be executed with RUN be grouped together.