Added LABEL to Dockerfile
As per best practices on Docker's site, since the MAINTAINER instruction
is being deprecated, and as the LABEL instruction is a lot more flexible
in terms of the key-value pairs it can use, I have used LABEL and added
all necessary information to it. LABEL also helps when the image is
queried with 'docker inspect' command.