Added Compose file for openssh-alpine
The Docker Compose file added uses the openssh-alpine image from the
'harshavardhanj' repository on Docker Hub. Else, it builds the image
using the Dockerfile that is in the same directory as the Compose file.
The Compose file is used to start 3 replicas of the openssh service on a
swarm. You can use it with the 'docker stack deploy' command. This file
can be tested on the local machine using the 'docker-compose up -d'
command.