Changes to Dockerfile
Have added Nano text editor to the list of packages to be installed at
build time. Have added ARG to help shorten some of the lines in the RUN
command. Have edited to LABEL tag to reflect the new package addition.
Have edited the file as per Docker best practices.