Added improved Dockerfile
Have made a lot of changes to the old Dockerfile. The ability to pass a
username and password as an environment variable during deployment of
the container has been added. Also, to further help deploy the server,
an 'entrypoint.sh' script has been included. This script will be called
by the ENTRYPOINT command in the Dockerfile. The script takes over from
then onwards and proceeds to configure the user-related settings for the
SSH server. For more information, please read the Dockerfile and its
corresponding 'entrypoint.sh' script file.