Major changes and additions
Have refined the Dockerfile further. Have added an 'entrypoint.sh'
script that will take care of setting up the SSH server as per the
values passed by the user. If none are passed, the defaults are used.
There is also an option to add an SSH public key. Using this, the user
can log in with key-based authentication.