Major changes and additions
Have refined the Dockerfile further. Have added an 'entrypoint.sh'
script that will take care of setting up the SSH server as per the
values passed by the user. If none are passed, the defaults are used.
There is also an option to add an SSH public key. Using this, the user
can log in with key-based authentication.

Due to the 'entrypoint.sh' script, the 'bash' shell had to be installed
as there were quite a few bashisms in the script which could not be run
using the 'sh' shell provided on the base image of Alpine Linux. This
resulted in the image size increasing from 6.41MB to 10.5MB.