Changes to Dockerfile and 'entrypoint.sh' script
Have removed the installation of 'bash' shell from the Dockerfile. This
was needed previously due to the 'entrypoint.sh' script which was taken
from the official Postgres image from Docker Hub. Now that the script
has been modified to work with '/bin/sh', there is no need for the
installation of 'bash'.

Have also added the 'exec "$@"' command at the end of the
'entrypoint.sh' script which will enable us to pass 'sh' as the command
to be run with the 'docker container run' command.