Added default user and password
Have added the default user that will be created as 'docker' and the
password of the default user will be 'docker'. The script reverts to the
deafault user and password when the USER and PASSWORD environment
variables are not set during 'docker container run ... -e USER=test -e
PASSWORD=test image', or in the Docker Compose file.