Changes to the if-elif-else conditional
In the previous version, if the USER variable provided on the command
line existed in the base distribution of Alpine Linux, there would be no
change of password, addition of '.ssh' folder in that user's home
directory, or modification of the 'config' file within the '.ssh'
directory.

Now, the if-elif-else conditional has been modified to accomodate the
passing of the USER variable which matches a pre-existing user on the
base system of Alpine Linux. The password of that existing user will be
changed as per the PASSWORD variable provided by the user, the '.ssh'
directory will be created in the home-directory of that user, which
already exists, and the 'config' file will be added and modified for SSH
login purposes.