Added missing command under 'root'
Have added a missing command under the if-conditional for the 'root'
user. The command changes 'PermitRootLogin no' to 'PermitRootLogin yes'
in the '/etc/ssh/sshd_config' file.

Have also added the same 'echo' commands from the 'entrypoint.sh' script
from the 'with-sh' branch. This command suggests that the user should
use SSH keys to login as 'root'.