Container is built from scratch image.
Have separated the image into two tags. The default image is built on
top of the scratch image. The busybox image is built on top of busybox.
The image built on scratch is much smaller than the image built on top
of busybox. Have created a folder to hold the files for the
busybox-based image.