Modified the Dockerfile to be independent of host
The Dockerfile has been modified so that it can be run on any system
without requiring the C file and the Makefile which were required
previously. These files were copied into the container for compilation.
Now, the source code is downloaded from GitHub and the compilation is
done. In essence, the ADD command has been removed.