Modified cloudbuild.yaml of docker-buildx
Have modified docker-buildx's cloudbuild.yaml in order to push the image
of docker-buildx:non-root to Docker Hub. The build step involving the
use of docker-buildx in order to build multi-arch images has been
commented out. Have done this because it's easier to build and push
large images from GCP. The upload speed of a residential connection
takes too long when uploading large images. In the next commit, if this
image works as intended, the build step which has been commented will be
included again.