Trying to run docker-buildx as non-root
Have added the following steps to the Dockerfile of docker-buildx:
  * Installation of 'su-exec' package to facilitate running commands
    as other users

Have changed the entrypoint.sh script to run 'docker-buildx` commands as
non-root user. First the script checks if the non-root user exists. If
the user exists, buildx is bootstrapped and all arguments are passed to
the buildx command which is run as the non-root user.