Changes to cloudbuild.yaml
Have commented out the buildx step and added the step where the non-root
image is built using docker. This is done because the previous non-root
image on Docker Hub needs to be replaced by the new image in order for
the changes made to the Dockerfile and entrypoint.sh files to take
effect. This way the new image can be tested using docker-buildx.