Trying to fix auth failure in buildx
I suspect that the insufficient auth error during the final stages of
pushing the image via buildx could be because buildx is running as
non-root user. Even though we are logged in to Docker Hub in a previous
step of Cloud Build, that access might not be propagating to the
non-root user. Therefore, a function 'dockerLogin' has been added to the
entrypoint.sh script, which takes the UserID and AccessToken files(which
must be present in the root directory since Cloud Build mounts it there)
and logs in using su-exec as non-root user. This may or may not work.
That remains to be seen.