buildx:non-root works now
Have changed the 'dockerfile.github.page' in the LABELS in the
Dockerfile for the 'buildx:non-root' image. Have added a few more steps
to the container-structure-test file. Have added a few more options to
speed up the build in cloudbuild.yaml. Have added a '--cache-from'
option to both 'docker build' and 'docker buildx' commands which will
help speed up the build if the layers have not changed. Have also
cleaned up the entrypoint.sh script. Finally, there's a '.gcloudignore'
file added to see if it helps preventing upload of all directories to
Cloud Build when the build is submitted.