Issue with build and buildx
Building and pushing with both 'docker build' and 'docker buildx'
results in the multi-arch images being overridden by the image built by
the 'docker build' command. This is because in order to push the image
built by the 'docker build' command, the image needs to be specified
under the 'images' section in cloudbuild.yaml. When this is done, at the
end of all the build steps, the image mentioned in the 'images' option
is pushed to Docker Hub. Since the images built by buildx are pushed
before the image built by 'docker build', the multi-arch images are
overridden. Therefore, to push multi-arch images, comment out the step
containing the 'docker build' command and the 'images' option.