Analysed with `dive` tool
Have analysed the `ddclient` image with `dive`. Based on the findings,
have found out that some repository information was being cached in the
'/var/cache/apk' directory. Have changed some of the commands in the RUN
directive in the Dockerfile.

The `apk update` command has been changed to `apk update -U --no-cache`.
The '-U' flag updates the cache, and the '--no-cache' option does not
use the local cache path. This should result in no cache being stored.
Finally, the virtual name '.ddclient-deps' has been removed as this
results in storing the data somewhere. As these deps are not going to be
removed, there is no necessity to register them under virtual packages.