Removed riscv
Have removed the stage of the Dockerfile which attempted to add support
for RISC in binfmt. The reason for this is that Moby does not support
building images with RISC yet. Therefore, building docker-buildx with
support for RISC is a moot point now. Until RISC builds are supported on
Moby, RISC support will not be added to this image.