Changes to Dockerfile
Have added pinned versions of Alpine Linux, and other packages or images
used in the Dockerfile. Same goes for the entrypoint.sh file which has
been improved. Now, the Dockerfile and entrypoint have been structured
such that any command passed to the builder will be same as passing
arguments to the 'docker' command. Now, 'docker build' will invoke the
'buildx' builder instead of the old builder. Have also added some tests
to the container-structure-test file. Finally, have added tags with
pinned version number corresponding to the version of docker installed
in the image.