Version number changes and others
Have changed the Golang version number to golang1.14-alpine3.12 and
Alpine Linux version to 3.12.0(for the second stage of the build). Have
added 'libc-dev' and 'libc6-compat' to the list of packages to be
installed during the first stage. This helps get rid of build errors for
some architectures. Have also segregated the files into folders named
after the version number of the 'pwndrop' release.